feat: add toast and notification for fox farming claims, also remove status page
