+++ src/components/Layout/Header/ActionCenter/components/Details/ClaimDetails.tsx
@@ -7,7 +7,7 @@ export const ClaimDetails = () => {
     <Stack gap={4}>
       <ButtonGroup width='full' size='sm'>
         <Button width='full' colorScheme='green'>
-          {translate('actionCenter.claim')}
+          {translate('actionCenter.claim.cta')}
         </Button>
       </ButtonGroup>
     </Stack>
